added in basic css for navigation and extracted it into a Header
