--- src/About.js
@@ -1,5 +1,0 @@
-const About = (props) => (
-    <h1> This is the About page </h1>
-)
-
-export default About
