+++ src/header/header.js
@@ -1,0 +1,9 @@
+import Navigation from "./navigation/navigation"
+
+const Header = () => (
+    <header>
+        <Navigation />
+    </header>
+)
+
+export default Header
