+++ src/header/navigation/navigation.css
@@ -1,0 +1,6 @@
+.navigation-list {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-evenly;
+    list-style: none;
+}
