--- src/navigation/navigation.js
@@ -1,11 +1,0 @@
-import { Link } from "react-router-dom";
-
-const Navigation = () => (
-    <nav>
-        <Link to="/">Home</Link>
-        <Link to="/about">About</Link>
-        <Link to="/Class">Class</Link>
-    </nav>
-)
-
-export default Navigation
